refactorizacion 3

+++ src/main/java/com/itsight/component/LoginListener.java
@@ -42,7 +42,7 @@ public class LoginListener implements ApplicationListener<InteractiveAuthenticat
             String username = usernameAndId[0], id = usernameAndId[1];
             session.setAttribute("id", Integer.parseInt(id));
             if(login.getAuthentication().getAuthorities().contains(new SimpleGrantedAuthority("ROLE_TRAINER"))){
-                Optional<String> optionalCodTrainer = Optional.ofNullable(trainerService.findCodigoTrainerById(Long.parseLong(id)));
+                Optional<String> optionalCodTrainer = Optional.ofNullable(trainerService.findCodigoTrainerById(Integer.parseInt(id)));
                 if(optionalCodTrainer.isPresent() && optionalCodTrainer.get().length()>0) {
                     session.setAttribute("codTrainer", optionalCodTrainer.get());
                 }
