+++ src/main/java/com/itsight/repository/DocumentoRepository.java
@@ -11,7 +11,7 @@ import java.util.List;
 import java.util.UUID;
 
 @Repository
-public interface DocumentoRepository extends JpaRepository<Documento, Long> {
+public interface DocumentoRepository extends JpaRepository<Documento, Integer> {
 
     @Override
     @EntityGraph(value = "documento", attributePaths = {})
@@ -27,15 +27,15 @@ public interface DocumentoRepository extends JpaRepository<Documento, Long> {
     List<Documento> findAllByNombreContainingIgnoreCase(String nombre);
 
     @EntityGraph(value = "documento", attributePaths = {})
-    Documento findById(Long id);
+    Documento findById(Integer id);
 
     @Query("SELECT D.nombre FROM Documento D WHERE D.id = ?1 AND D.uuid = ?2")
-    String findNombreByIdAndUuid(Long id, UUID uuid);
+    String findNombreByIdAndUuid(Integer id, UUID uuid);
 
     @EntityGraph(value = "documento", attributePaths = {})
     List<Documento> findAllByNombreContainingIgnoreCaseAndFlagActivo(String nombre, Boolean flagActivo);
 
     @Modifying
     @Query(value = "UPDATE Documento D SET D.flagActivo =?2 WHERE D.id = ?1")
-    void updateFlagActivoById(Long id, boolean flagActivo);
+    void updateFlagActivoById(Integer id, boolean flagActivo);
 }
