+++ src/main/java/com/itsight/repository/RuConsolidadoRepository.java
@@ -8,13 +8,13 @@ import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface RuConsolidadoRepository extends JpaRepository<RuConsolidado, Long> {
+public interface RuConsolidadoRepository extends JpaRepository<RuConsolidado, Integer> {
 
     @EntityGraph(value = "ruConsolidado.rutina")
-    RuConsolidado getById(Long id);
+    RuConsolidado getById(Integer id);
 
     @Modifying
     @Query("UPDATE RuConsolidado R SET R.matrizMejoraVelocidades = ?2 WHERE R.id = ?1")
-    void updateMatrizMejoraVelocidades(Long rutinaId, String mVz);
+    void updateMatrizMejoraVelocidades(Integer rutinaId, String mVz);
 
 }
